Refactor column alias logic
